ml/requirements.txt
pandas
scikit-learn
python-dotenv
requests